USA Credit Card Default Info Power BI project
In this project, I use data from credit card defaults in Taiwan from 2005. The US states are simply added in order to add a geographical element to my dashboard but are a fictional addition to the data. This dashboard has very basic functionality and the main use cases are to see:

1. What education level the defaulter has.
2. What sex the defaulter is.
3. In which state the defaulter lives.
4. What marital status the defaulter has.